refactor: Split large files to fix ESLint max-lines violations
Split CanvasController.path.test.ts (1237 lines) into 6 modular test files:
- path.setup.ts: Shared test setup and mocks
- path.basic.test.ts: Basic path operations
- path.arcs.test.ts: Arc operations
- path.curves.test.ts: Bezier curves
- path.shapes.test.ts: Shapes (roundRect, clip)
- path.styling.test.ts: Styling operations
- path.tracking.test.ts: Path2D tracking, hit testing

Split canvasLuaWrapper.ts (767 lines) into modular Lua code files:
- canvasLuaCode/core.ts: Core canvas functions
- canvasLuaCode/path.ts: Path API and pixel manipulation
- canvasLuaCode/styling.ts: Line styles, gradients, shadows
- canvasLuaCode/text.ts: Text alignment and draw_label
- canvasLuaCode/input.ts: Input handling and key constants
