+++ packages/lua-runtime/src/canvasLuaCode/index.ts
@@ -1,0 +1,10 @@
+/**
+ * Canvas Lua code module exports.
+ * Each module contains a section of the Lua code for the canvas API.
+ */
+
+export { canvasLuaCoreCode } from './core'
+export { canvasLuaPathCode } from './path'
+export { canvasLuaStylingCode } from './styling'
+export { canvasLuaTextCode } from './text'
+export { canvasLuaInputCode } from './input'
