+++ packages/lua-runtime/src/canvasLuaCode/input.ts
@@ -1,0 +1,140 @@
+/**
+ * Input Lua code - keyboard, mouse input and key constants.
+ */
+
+export const canvasLuaInputCode = `
+    -- Helper to normalize key names
+    local function normalize_key(key)
+      if type(key) ~= 'string' then return key end
+      -- Single letter keys
+      if #key == 1 and key:match('%a') then
+        return 'Key' .. key:upper()
+      end
+      -- Arrow keys
+      local arrows = { up = 'ArrowUp', down = 'ArrowDown', left = 'ArrowLeft', right = 'ArrowRight' }
+      if arrows[key:lower()] then
+        return arrows[key:lower()]
+      end
+      -- Space key
+      if key:lower() == 'space' or key == ' ' then
+        return 'Space'
+      end
+      -- Common keys
+      local common = {
+        enter = 'Enter', escape = 'Escape', esc = 'Escape',
+        tab = 'Tab', shift = 'ShiftLeft', ctrl = 'ControlLeft',
+        alt = 'AltLeft', backspace = 'Backspace'
+      }
+      if common[key:lower()] then
+        return common[key:lower()]
+      end
+      return key
+    end
+
+    -- Keyboard input
+    function _canvas.is_key_down(key)
+      return __canvas_isKeyDown(normalize_key(key))
+    end
+
+    function _canvas.is_key_pressed(key)
+      return __canvas_isKeyPressed(normalize_key(key))
+    end
+
+    -- Helper to convert JS array proxy to plain Lua table
+    -- This ensures proper Lua errors with line numbers instead of JS TypeErrors
+    local function to_lua_array(js_array)
+      local t = {}
+      for i = 1, #js_array do
+        t[i] = js_array[i]
+      end
+      return t
+    end
+
+    function _canvas.get_keys_down()
+      return to_lua_array(__canvas_getKeysDown())
+    end
+
+    function _canvas.get_keys_pressed()
+      return to_lua_array(__canvas_getKeysPressed())
+    end
+
+    -- Mouse input
+    function _canvas.get_mouse_x()
+      return __canvas_getMouseX()
+    end
+
+    function _canvas.get_mouse_y()
+      return __canvas_getMouseY()
+    end
+
+    function _canvas.is_mouse_down(button)
+      return __canvas_isMouseDown(button)
+    end
+
+    function _canvas.is_mouse_pressed(button)
+      return __canvas_isMousePressed(button)
+    end
+
+    -- Key constants for discoverability
+    _canvas.keys = {
+      -- Letters
+      A = 'KeyA', B = 'KeyB', C = 'KeyC', D = 'KeyD', E = 'KeyE',
+      F = 'KeyF', G = 'KeyG', H = 'KeyH', I = 'KeyI', J = 'KeyJ',
+      K = 'KeyK', L = 'KeyL', M = 'KeyM', N = 'KeyN', O = 'KeyO',
+      P = 'KeyP', Q = 'KeyQ', R = 'KeyR', S = 'KeyS', T = 'KeyT',
+      U = 'KeyU', V = 'KeyV', W = 'KeyW', X = 'KeyX', Y = 'KeyY', Z = 'KeyZ',
+
+      -- Number row
+      ['0'] = 'Digit0', ['1'] = 'Digit1', ['2'] = 'Digit2', ['3'] = 'Digit3',
+      ['4'] = 'Digit4', ['5'] = 'Digit5', ['6'] = 'Digit6', ['7'] = 'Digit7',
+      ['8'] = 'Digit8', ['9'] = 'Digit9',
+      DIGIT_0 = 'Digit0', DIGIT_1 = 'Digit1', DIGIT_2 = 'Digit2', DIGIT_3 = 'Digit3',
+      DIGIT_4 = 'Digit4', DIGIT_5 = 'Digit5', DIGIT_6 = 'Digit6', DIGIT_7 = 'Digit7',
+      DIGIT_8 = 'Digit8', DIGIT_9 = 'Digit9',
+
+      -- Arrow keys
+      UP = 'ArrowUp', DOWN = 'ArrowDown', LEFT = 'ArrowLeft', RIGHT = 'ArrowRight',
+      ARROW_UP = 'ArrowUp', ARROW_DOWN = 'ArrowDown', ARROW_LEFT = 'ArrowLeft', ARROW_RIGHT = 'ArrowRight',
+
+      -- Function keys
+      F1 = 'F1', F2 = 'F2', F3 = 'F3', F4 = 'F4', F5 = 'F5', F6 = 'F6',
+      F7 = 'F7', F8 = 'F8', F9 = 'F9', F10 = 'F10', F11 = 'F11', F12 = 'F12',
+
+      -- Modifier keys
+      SHIFT = 'ShiftLeft', SHIFT_LEFT = 'ShiftLeft', SHIFT_RIGHT = 'ShiftRight',
+      CTRL = 'ControlLeft', CTRL_LEFT = 'ControlLeft', CTRL_RIGHT = 'ControlRight',
+      CONTROL = 'ControlLeft', CONTROL_LEFT = 'ControlLeft', CONTROL_RIGHT = 'ControlRight',
+      ALT = 'AltLeft', ALT_LEFT = 'AltLeft', ALT_RIGHT = 'AltRight',
+      META = 'MetaLeft', META_LEFT = 'MetaLeft', META_RIGHT = 'MetaRight',
+      CAPS_LOCK = 'CapsLock',
+
+      -- Special keys
+      SPACE = 'Space', ENTER = 'Enter', ESCAPE = 'Escape', TAB = 'Tab',
+      BACKSPACE = 'Backspace', DELETE = 'Delete', INSERT = 'Insert',
+      HOME = 'Home', END = 'End', PAGE_UP = 'PageUp', PAGE_DOWN = 'PageDown',
+      PRINT_SCREEN = 'PrintScreen', SCROLL_LOCK = 'ScrollLock', PAUSE = 'Pause',
+      NUM_LOCK = 'NumLock',
+
+      -- Numpad keys
+      NUMPAD_0 = 'Numpad0', NUMPAD_1 = 'Numpad1', NUMPAD_2 = 'Numpad2', NUMPAD_3 = 'Numpad3',
+      NUMPAD_4 = 'Numpad4', NUMPAD_5 = 'Numpad5', NUMPAD_6 = 'Numpad6', NUMPAD_7 = 'Numpad7',
+      NUMPAD_8 = 'Numpad8', NUMPAD_9 = 'Numpad9',
+      NUMPAD_ADD = 'NumpadAdd', NUMPAD_SUBTRACT = 'NumpadSubtract',
+      NUMPAD_MULTIPLY = 'NumpadMultiply', NUMPAD_DIVIDE = 'NumpadDivide',
+      NUMPAD_DECIMAL = 'NumpadDecimal', NUMPAD_ENTER = 'NumpadEnter',
+
+      -- Punctuation and symbols
+      MINUS = 'Minus', EQUAL = 'Equal', BRACKET_LEFT = 'BracketLeft', BRACKET_RIGHT = 'BracketRight',
+      BACKSLASH = 'Backslash', SEMICOLON = 'Semicolon', QUOTE = 'Quote',
+      BACKQUOTE = 'Backquote', COMMA = 'Comma', PERIOD = 'Period', SLASH = 'Slash',
+
+      -- Context menu key
+      CONTEXT_MENU = 'ContextMenu',
+    }
+
+    -- Register canvas as a module so require('canvas') works
+    -- Note: canvas is NOT a global - it must be accessed via require('canvas')
+    package.preload['canvas'] = function()
+      return _canvas
+    end
+`
